pyproject.toml
[tool.poetry]
name = "django-repo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
ollama = "^0.4.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
